feat (*): clean up ingredient hooks.
This commit changes how the app handles fetching data. Each domain
object that has a view (ie, Ingredient/Recipe) should have a
corresponing hook for loading the entire set of data (we want to load
the entire set so it is available and can use a selctor for individual
pieces).

This hook will check the store to see if the value we want is already
populated. If it is, it will do nothing, allowing the component to use
the value already in the store. If it is not available, it will fetch
the data and populate the store for us -- this will cause an update to
the observer and a re-render of the component.